Moved oauth related stuff into separate file
